+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 1.5 branch
 ### 1.5.4
 * updated for `php-tools` 1.8;
-* `NotReadableImageException` no longer exists, replaced with a generic `LogicException`;
+* `NotReadableImageException` and `UnsupportedImageTypeException` exceptions no longer exist, replaced with a generic
+  `LogicException`;
 * no longer uses the `Exceptionist` class.
 
 ### 1.5.3